netty server can connect to socket client

+++ src/main/java/com/lab409/socket/demoServer/mapper/SensorMapper.java
@@ -1,6 +1,6 @@
 package com.lab409.socket.demoServer.mapper;
 
-import com.lab409.socket.demoServer.entity.Sensor;
+import com.lab409.socket.demoServer.model.Sensor;
 import org.apache.ibatis.annotations.*;
 
 import java.util.List;
@@ -10,7 +10,7 @@ public interface SensorMapper {
    List<Sensor> getAll();
 
    @Select("SELECT * FROM sensors WHERE id=#{id}")
-   Sensor getOne(Long id);
+   Sensor getOne(Integer id);
 
    @Insert("INSERT INTO sensors(id, type, descr, host, port, msg, state) VALUES(#{id}, #{type}, #{descr}, #{host}, #{port}, #{msg}, #{state})")
    void insert(Sensor sensor);
@@ -19,5 +19,5 @@ public interface SensorMapper {
     void update(Sensor sensor);
 
    @Delete("DELETE FROM sensors WHERE id=#{id}")
-    void delete(Long id);
+    void delete(Integer id);
 }
